+++ starminder/core/forms.py
@@ -3,15 +3,24 @@ from django import forms
 from starminder.core.models import UserProfile
 
 
+def _coerce(value: str) -> bool:
+    return value == "True"
+
+
 class UserProfileConfigForm(forms.ModelForm):
-    INCLUDE_ARCHIVED_CHOICES = [
+    INCLUDE_CHOICES = [
         (True, "Include"),
         (False, "Don't include"),
     ]
 
     include_archived = forms.TypedChoiceField(
-        choices=INCLUDE_ARCHIVED_CHOICES,
-        coerce=lambda x: x == "True",
+        choices=INCLUDE_CHOICES,
+        coerce=_coerce,
+        widget=forms.Select(),
+    )
+    include_own = forms.TypedChoiceField(
+        choices=INCLUDE_CHOICES,
+        coerce=_coerce,
         widget=forms.Select(),
     )
 
@@ -23,6 +32,7 @@ class UserProfileConfigForm(forms.ModelForm):
             "day_of_week",
             "hour_of_day",
             "include_archived",
+            "include_own",
         ]
         widgets = {
             "reminder_email": forms.EmailInput(),
@@ -36,4 +46,5 @@ class UserProfileConfigForm(forms.ModelForm):
             "day_of_week": "Day of week",
             "hour_of_day": "Hour of day (0-23)",
             "include_archived": "Archived repositories",
+            "include_own": "Own repositories",
         }